Move thread pool to separate hpp and cpp file
